fixed poc ruleset
